@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.44+2
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -34,7 +34,7 @@ dependencies:
   firebase_messaging: ^11.1.0
   http: ^0.13.3
   image_picker: ^0.8.4+1
-  image_cropper: ^1.3.1
+  image_cropper: ^2.0.2
   numeric_keyboard: ^1.0.0
   search_choices : ^2.0.19
   shared_preferences: ^2.0.7
