@@ -20,7 +20,7 @@ class ShareUrlPage extends StatefulWidget {
 
 class _ShareUrlPageState extends State<ShareUrlPage> {
   var lowPurchase = new MoneyMaskedTextController(initialValue: 0, decimalSeparator: ',', thousandSeparator: '.',  leftSymbol: '\$ ', );
-  double _total;
+  double? _total;
   int _statusButtonNetworks = 0;
   bool _statusButtonReady = false;
 
@@ -247,15 +247,15 @@ class _ShareUrlPageState extends State<ShareUrlPage> {
       priceDouble = double.parse(item['data'].price);
       priceDouble *= item['quantity'];
       if(item['data'].coin == 0 && myProvider.coinUsers == 1)
-        _total+=(priceDouble * varRate);
+        _total= _total! + (priceDouble * varRate);
       else if(item['data'].coin == 1 && myProvider.coinUsers == 0)
-        _total+=(priceDouble / varRate);
+        _total= _total! + (priceDouble / varRate);
       else
-        _total+=(priceDouble);
+        _total= _total! + (priceDouble);
 
     }
 
-    lowPurchase.updateValue(_total);
+    lowPurchase.updateValue(_total!);
 
     return "${lowPurchase.text}";
   }  
