@@ -26,7 +26,7 @@ class _RegisterPageState extends State<RegisterPage> {
   final FocusNode _emailFocus = FocusNode();  
   final FocusNode _passwordFocus = FocusNode();
   final FocusNode _passwordConfirmFocus = FocusNode();
-  String _name, _address, _phone, _email, _password, _passwordConfirm, _messageError;
+  String? _name, _address, _phone, _email, _password, _passwordConfirm, _messageError='';
   bool passwordVisible = true, _statusError = false;
   var jsonResponse;
 
@@ -177,7 +177,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
               ),
               validator: _validateEmail,
-              onSaved: (String? value) => _email = value.toLowerCase().trim(),
+              onSaved: (String? value) => _email = value!.toLowerCase().trim(),
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
               style: TextStyle(
@@ -285,7 +285,7 @@ class _RegisterPageState extends State<RegisterPage> {
               child: Padding(
                 padding: const EdgeInsets.fromLTRB(10.0, 15.0, 0.0, 20.0),
                 child: AutoSizeText(
-                  _messageError == null? '' : _messageError,
+                  _messageError!.isEmpty? '' : _messageError!,
                   style: TextStyle(
                     color: Colors.red,
                     fontFamily: 'MontserratSemiBold',
@@ -438,7 +438,7 @@ class _RegisterPageState extends State<RegisterPage> {
             myProvider.statusButtonMenu = false;
             myProvider.clickButtonMenu = 0;
             myProvider.dataShipping = [];
-            myProvider.getTokenFCM = null;
+            myProvider.getTokenFCM = '';
           } else if(jsonResponse['errors'] != null){
 
             setState(() {
@@ -511,12 +511,12 @@ class _RegisterPageState extends State<RegisterPage> {
     );
   }
 
-  String _validateName(String? value) {
+  String? _validateName(String? value) {
     // This is just a regular expression for name
     String p = '[a-zA-Z]';
     RegExp regExp = new RegExp(p);
 
-    if (value.isNotEmpty && regExp.hasMatch(value) && value.length >=3) {
+    if (value!.isNotEmpty && regExp.hasMatch(value) && value.length >=3) {
       // So, the name is valid
       return null;
     }
@@ -525,9 +525,9 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese nombre y apellido válido';
   }
 
-  String _validateAddress(String? value) {
+  String? _validateAddress(String? value) {
 
-    if (value.length >=3) {
+    if (value!.length >=3) {
       // So, the address is valid
       return null;
     }
@@ -536,13 +536,13 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese una dirección válido';
   }
 
-  String _validatePhone(String? value) {
+  String? _validatePhone(String? value) {
     // This is just a regular expression for phone
     //String p = r'^(?:(\+)58|0)(?:2(?:12|4[0-9]|5[1-9]|6[0-9]|7[0-8]|8[1-35-8]|9[1-5]|3[45789])|4(?:1[246]|2[46]))\d{7}$';
     String p = r'^(0414|0424|0412|0416|0426)[0-9]{7}$';
     RegExp regExp = new RegExp(p);
 
-    if (value.isNotEmpty && regExp.hasMatch(value) && value.length >=9) {
+    if (value!.isNotEmpty && regExp.hasMatch(value) && value.length >=9) {
       // So, the phone is valid
       return null;
     }
@@ -551,8 +551,8 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese un número de teléfono válido';
   }
 
-  String _validateEmail(String? value) {
-    value = value.trim().toLowerCase();
+  String? _validateEmail(String? value) {
+    value = value!.trim().toLowerCase();
     // This is just a regular expression for email addresses
     String p = "[a-zA-Z0-9\+\.\_\%\-\+]{1,256}" +
         "\\@" +
@@ -572,9 +572,9 @@ class _RegisterPageState extends State<RegisterPage> {
   }
 
 
-  String _validatePassword(String? value) {
+  String? _validatePassword(String? value) {
     String errorValidate = 'La contraseña es inválida, debe tener:';
-    if (value.isEmpty) {
+    if (value!.isEmpty) {
       // The form is empty
       return 'Ingrese una contraseña válido';
     }
@@ -615,8 +615,8 @@ class _RegisterPageState extends State<RegisterPage> {
 
   }
 
-  String _validatePasswordConfirm(String? value) {
-    if(value.isNotEmpty){
+  String? _validatePasswordConfirm(String? value) {
+    if(value!.isNotEmpty){
       if (_password == value){
         return null;
       }
